+++ algosdk/auction.py
@@ -16,7 +16,7 @@ class Bid:
         auction_id (int): auction ID
     """
 
-    def __init__(self, bidder: str, bid_currency: int, max_price: int, bid_id: int, auction_key: str, auction_id: int):
+    def __init__(self, bidder: str, bid_currency: int, max_price: int, bid_id: int, auction_key: str, auction_id: int) -> None:
         self.bidder = bidder
         self.bid_currency = bid_currency
         self.max_price = max_price
@@ -24,7 +24,7 @@ class Bid:
         self.auction_key = auction_key
         self.auction_id = auction_id
 
-    def dictify(self):
+    def dictify(self) -> dict:
         d = {
             "aid": self.auction_id,
             "auc": util.decode_address(self.auction_key),
@@ -35,7 +35,7 @@ class Bid:
         }
         return d
 
-    def sign(self, private_key: str):
+    def sign(self, private_key: str) -> "SignedBid":
         """
         Sign a bid.
 
@@ -55,11 +55,11 @@ class Bid:
         return signed
 
     @staticmethod
-    def undictify(d: dict):
+    def undictify(d: dict) -> "Bid":
         return Bid(util.encode_address(d["bidder"]), d["cur"], d["price"],
                    d["id"], util.encode_address(d["auc"]), d["aid"])
 
-    def __eq__(self, other: "Bid"):
+    def __eq__(self, other: "Bid") -> bool:
         if not isinstance(other, Bid):
             return False
         return (self.bidder == other.bidder and
@@ -78,11 +78,11 @@ class SignedBid:
         bid (Bid): bid that was signed
         signature (str): the signature of the bidder
     """
-    def __init__(self, bid: Bid, signature: str):
+    def __init__(self, bid: Bid, signature: str) -> None:
         self.bid = bid
         self.signature = signature
 
-    def dictify(self):
+    def dictify(self) -> dict:
         d = {
             "bid": self.bid.dictify(),
             "sig": base64.b64decode(self.signature)
@@ -90,11 +90,11 @@ class SignedBid:
         return d
 
     @staticmethod
-    def undictify(d: dict):
+    def undictify(d: dict) -> "SignedBid":
         return SignedBid(Bid.undictify(d["bid"]),
                          base64.b64encode(d["sig"]).decode())
 
-    def __eq__(self, other: "SignedBid"):
+    def __eq__(self, other: "SignedBid") -> bool:
         if not isinstance(other, SignedBid):
             return False
         return (self.bid == other.bid and
@@ -109,11 +109,11 @@ class NoteField:
         signed_bid (SignedBid): bid with signature of bidder
         note_field_type (str): the type of note; see constants for possible types
     """
-    def __init__(self, signed_bid: SignedBid, note_field_type: str):
+    def __init__(self, signed_bid: SignedBid, note_field_type: str) -> None:
         self.signed_bid = signed_bid
         self.note_field_type = note_field_type
 
-    def dictify(self):
+    def dictify(self) -> dict:
         d = {
             "b": self.signed_bid.dictify(),
             "t": self.note_field_type
@@ -121,10 +121,10 @@ class NoteField:
         return d
 
     @staticmethod
-    def undictify(d: dict):
+    def undictify(d: dict) -> "NoteField":
         return NoteField(SignedBid.undictify(d["b"]), d["t"])
 
-    def __eq__(self, other: "NoteField"):
+    def __eq__(self, other: "NoteField") -> bool:
         if not isinstance(other, NoteField):
             return False
         return (self.signed_bid == other.signed_bid and