+++ algosdk/util.py
@@ -110,8 +110,7 @@ def msgpack_encode(obj):
     Encode the object using canonical msgpack.
 
     Args:
-        obj (Transaction, SignedTransaction, MultisigTransaction, Multisig,\
-            Bid, or SignedBid): object to be encoded
+        obj (Object): object to be encoded
 
     Returns:
         str: msgpack encoded object
@@ -151,10 +150,11 @@ def msgpack_decode(enc):
         enc (str): string to be decoded
 
     Returns:
-        Transaction, SignedTransaction, Multisig, Bid, or SignedBid:\
-            decoded object
+        Object: decoded object
     """
-    decoded = msgpack.unpackb(base64.b64decode(enc), raw=False)
+    decoded = enc
+    if not isinstance(enc, dict):
+        decoded = msgpack.unpackb(base64.b64decode(enc), raw=False)
     if "type" in decoded:
         return transaction.Transaction.undictify(decoded)
     if "l" in decoded:
@@ -163,8 +163,10 @@ def msgpack_decode(enc):
         return transaction.MultisigTransaction.undictify(decoded)
     if "lsig" in decoded:
         return transaction.LogicSigTransaction.undictify(decoded)
-    if "txn" in decoded:
+    if "sig" in decoded:
         return transaction.SignedTransaction.undictify(decoded)
+    if "txn" in decoded:
+        return transaction.Transaction.undictify(decoded["txn"])
     if "subsig" in decoded:
         return transaction.Multisig.undictify(decoded)
     if "txlist" in decoded:
@@ -284,13 +286,12 @@ def checksum(data):
     return chksum.finalize()
 
 
-def write_to_file(txns, path, overwrite=True):
+def write_to_file(objs, path, overwrite=True):
     """
-    Write signed or unsigned transactions to a file.
+    Write objects to a file.
 
     Args:
-        txns (Transaction[], SignedTransaction[], or MultisigTransaction[]):\
-            can be a mix of the three
+        objs (Object[]): list of encodable objects
         path (str): file to write to
         overwrite (bool): whether or not to overwrite what's already in the
             file; if False, transactions will be appended to the file
@@ -305,40 +306,31 @@ def write_to_file(txns, path, overwrite=True):
     else:
         f = open(path, "ab")
 
-    for txn in txns:
-        if isinstance(txn, transaction.Transaction):
-            enc = msgpack.packb({"txn": txn.dictify()}, use_bin_type=True)
-            f.write(enc)
+    for obj in objs:
+        if isinstance(obj, transaction.Transaction):
+            f.write(base64.b64decode(msgpack_encode({"txn": obj.dictify()})))
         else:
-            enc = msgpack.packb(txn.dictify(), use_bin_type=True)
-            f.write(enc)
+            f.write(base64.b64decode(msgpack_encode(obj)))
+
     f.close()
     return True
 
 
 def read_from_file(path):
     """
-    Retrieve signed or unsigned transactions from a file.
+    Retrieve encoded objects from a file.
 
     Args:
         path (str): file to read from
 
     Returns:
-        Transaction[], SignedTransaction[], or MultisigTransaction[]:\
-            can be a mix of the three
+        Object[]: list of objects
     """
 
     f = open(path, "rb")
-    txns = []
+    objs = []
     unp = msgpack.Unpacker(f, raw=False)
-    for txn in unp:
-        if "msig" in txn:
-            txns.append(transaction.MultisigTransaction.undictify(txn))
-        elif "sig" in txn:
-            txns.append(transaction.SignedTransaction.undictify(txn))
-        elif "lsig" in txn:
-            txns.append(transaction.LogicSigTransaction.undictify(txn))
-        elif "txn" in txn:
-            txns.append(transaction.Transaction.undictify(txn["txn"]))
+    for obj in unp:
+        objs.append(msgpack_decode(obj))
     f.close()
-    return txns
+    return objs